Make DBHandler interface more abstract
The DBHandler interface assumed that openStream, readStream and
closeStream used and returned ifstream objects. This resulted in the
interface not being generic enough and compelled the user to store the
database in a file.

Remove the DBHandler's interface dependency on ifstreams. Instead assume
that openStream accepts a uint64_t as the stream's identifier. Then
readStream accepts the identifier of an open stream to return the actual
data. Likewise closeStream accepts the identifier of an opened stream
and closes it.